abandon use of OE heap
