+++ include/myst/tcall.h
@@ -18,8 +18,6 @@
 typedef enum myst_tcall_number
 {
     MYST_TCALL_RANDOM = 2048,
-    MYST_TCALL_ALLOCATE = 2049,
-    MYST_TCALL_DEALLOCATE = 2050,
     MYST_TCALL_VSNPRINTF = 2051,
     MYST_TCALL_WRITE_CONSOLE = 2052,
     MYST_TCALL_GEN_CREDS = 2053,
