+++ kernel/libc.c
@@ -427,259 +427,6 @@ int putchar(int c)
 /*
 **==============================================================================
 **
-** <stdlib.h>
-**
-**==============================================================================
-*/
-
-#define MAX_BACKTRACE_ADDRS 16
-
-static long _tcall_allocate(
-    void* ptr,
-    size_t alignment,
-    size_t size,
-    int clear,
-    void** new_ptr)
-{
-    long params[6];
-    params[0] = (long)ptr;
-    params[1] = (long)alignment;
-    params[2] = (long)size;
-    params[3] = (long)clear;
-    params[4] = (long)new_ptr;
-
-    return myst_tcall(MYST_TCALL_ALLOCATE, params);
-}
-
-static long _tcall_deallocate(void* ptr)
-{
-    long params[6];
-    params[0] = (long)ptr;
-
-    return myst_tcall(MYST_TCALL_DEALLOCATE, params);
-}
-#ifdef MYST_ENABLE_LEAK_CHECKER
-static myst_list_t _list;
-static myst_spinlock_t _lock = MYST_SPINLOCK_INITIALIZER;
-#endif
-
-static myst_malloc_stats_t _malloc_stats;
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-typedef struct node
-{
-    myst_list_node_t base;
-    void* ptr;
-    size_t size;
-    void* addrs[MAX_BACKTRACE_ADDRS];
-    size_t num_addrs;
-} node_t;
-#endif
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-static node_t* _new_node(void* ptr, size_t size)
-{
-    node_t* p;
-
-    if (_tcall_allocate(NULL, 0, sizeof(node_t), 0, (void**)&p) != 0 || !p)
-        return NULL;
-
-    p->ptr = ptr;
-    p->size = size;
-    p->num_addrs = myst_backtrace(p->addrs, MYST_COUNTOF(p->addrs));
-
-    return p;
-}
-#endif
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-static int _add_node(void* ptr, size_t size)
-{
-    node_t* node;
-
-    if (!(node = _new_node(ptr, size)))
-        return -1;
-
-    myst_spin_lock(&_lock);
-    {
-        myst_list_append(&_list, (myst_list_node_t*)node);
-        _malloc_stats.usage += size;
-
-        if (_malloc_stats.usage > _malloc_stats.peak_usage)
-            _malloc_stats.peak_usage = _malloc_stats.usage;
-    }
-    myst_spin_unlock(&_lock);
-
-    return 0;
-}
-#endif
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-static int _remove_node(void* ptr)
-{
-    node_t* node = NULL;
-
-    myst_spin_lock(&_lock);
-    {
-        for (myst_list_node_t* p = _list.head; p; p = p->next)
-        {
-            node_t* tmp = (node_t*)p;
-
-            if (tmp->ptr == ptr)
-            {
-                node = tmp;
-                _malloc_stats.usage -= node->size;
-                myst_list_remove(&_list, p);
-                break;
-            }
-        }
-    }
-    myst_spin_unlock(&_lock);
-
-    if (node)
-    {
-        _tcall_deallocate(node);
-        return 0;
-    }
-
-    return -1;
-}
-#endif
-
-MYST_WEAK
-void* malloc(size_t size)
-{
-    void* p = NULL;
-
-    if (_tcall_allocate(NULL, 0, size, 0, &p) != 0 || !p)
-        return NULL;
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    if (_add_node(p, size) != 0)
-        myst_panic("unexpected");
-#endif
-
-    return p;
-}
-
-MYST_WEAK
-void* calloc(size_t nmemb, size_t size)
-{
-    void* p = NULL;
-    size_t n = nmemb * size;
-
-    if (_tcall_allocate(NULL, 0, n, 1, &p) != 0 || !p)
-        return NULL;
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    if (_add_node(p, n) != 0)
-        myst_panic("unexpected");
-#endif
-
-    return p;
-}
-
-MYST_WEAK
-void* realloc(void* ptr, size_t size)
-{
-    void* p = NULL;
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    if (ptr && _remove_node(ptr) != 0)
-        myst_panic("unexpected");
-#endif
-
-    if (_tcall_allocate(ptr, 0, size, 0, &p) != 0 || !p)
-        return NULL;
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    if (_add_node(p, size) != 0)
-        myst_panic("unexpected");
-#endif
-
-    return p;
-}
-
-MYST_WEAK
-void* memalign(size_t alignment, size_t size)
-{
-    void* p = NULL;
-
-    if (_tcall_allocate(NULL, alignment, size, 0, &p) != 0 || !p)
-        return NULL;
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    if (_add_node(p, size) != 0)
-        myst_panic("unexpected");
-#endif
-
-    return p;
-}
-
-MYST_WEAK
-void free(void* ptr)
-{
-    if (_tcall_deallocate(ptr) != 0)
-        myst_panic("unexpected");
-
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    if (ptr && _remove_node(ptr) != 0)
-        myst_panic("unexpected");
-#endif
-}
-
-int myst_find_leaks(void)
-{
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    int ret = 0;
-
-    for (myst_list_node_t* p = _list.head; p; p = p->next)
-    {
-        node_t* node = (node_t*)p;
-
-        myst_eprintf(
-            "*** kernel leak: ptr=%p size=%zu\n", node->ptr, node->size);
-        myst_dump_backtrace(node->addrs, node->num_addrs);
-        myst_eprintf("\n");
-        ret = -1;
-    }
-
-    if (_malloc_stats.usage != 0)
-    {
-        myst_eprintf(
-            "*** kernel: memory still in use: %zu\n", _malloc_stats.usage);
-        ret = -1;
-    }
-
-    return ret;
-#else
-    return 0;
-#endif
-}
-
-int myst_get_malloc_stats(myst_malloc_stats_t* stats)
-{
-#ifdef MYST_ENABLE_LEAK_CHECKER
-    {
-        if (!stats)
-            return -EINVAL;
-
-        myst_spin_lock(&_lock);
-        *stats = _malloc_stats;
-        myst_spin_unlock(&_lock);
-
-        return 0;
-    }
-#else
-    (void)stats;
-    (void)_malloc_stats;
-    return -ENOTSUP;
-#endif
-}
-
-/*
-**==============================================================================
-**
 ** <fcntl.h>
 **
 **==============================================================================
