+++ tools/myst/enc/enc.c
@@ -833,6 +833,10 @@ OE_SET_ENCLAVE_SGX(
     1,        /* ProductID */
     1,        /* SecurityVersion */
     true,     /* Debug */
+#if 1
+    16,       /* NumHeapPages */
+#else
     8 * 4096, /* NumHeapPages */
+#endif
     32,       /* NumStackPages */
     16);      /* NumTCS */
