+++ tools/myst/host/exec.h
@@ -14,7 +14,6 @@ int exec_launch_enclave(
     uint32_t flags,
     const char* argv[],
     const char* envp[],
-    struct myst_options* options,
-    oe_sgx_enclave_properties_t* enclave_properties);
+    struct myst_options* options);
 
 #endif /* _MYST_HOST_EXEC_H */
